# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leads CRM API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/v1/leads": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "List leads",
                "description": "Returns a paginated list of leads based on the user's role with filtering support.",
                "operationId": "d660caaedd473f1082c49dca5634bb53",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by lead status (same values as web)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Follow-up"
                        }
                    },
                    {
                        "name": "assigned_to",
                        "in": "query",
                        "description": "Filter by assigned sales user ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter leads created from this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter leads created until this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-31"
                        }
                    },
                    {
                        "name": "lead_id",
                        "in": "query",
                        "description": "Filter by exact lead ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lead"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/leads/{id}": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Get lead details",
                "description": "Returns lead details with last follow-up and permissions",
                "operationId": "getLeadDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lead ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "lead": {
                                                    "$ref": "#/components/schemas/Lead"
                                                },
                                                "last_follow_up": {
                                                    "$ref": "#/components/schemas/LastFollowUp"
                                                },
                                                "permissions": {
                                                    "$ref": "#/components/schemas/LeadPermissions"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lead not found or not visible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/leads/{id}/move-forward": {
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Move lead forward (change status)",
                "operationId": "moveLeadForward",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lead ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "description": "Idempotency key to prevent duplicate execution",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "3f2c8c8e-1a34-4a0e-9c7b-0f2d2e7c4b91"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_id",
                                    "comment"
                                ],
                                "properties": {
                                    "status_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "assigned_to": {
                                        "type": "integer",
                                        "example": 13,
                                        "nullable": true
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Client asked to follow up next week"
                                    },
                                    "reminder_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-01-10T10:30:00Z",
                                        "nullable": true
                                    },
                                    "meeting_range": {
                                        "type": "string",
                                        "example": "2026-01-10 10:00 - 2026-01-10 10:30",
                                        "nullable": true
                                    },
                                    "client_unit": {
                                        "type": "string",
                                        "example": "A-1203",
                                        "nullable": true
                                    },
                                    "client_price": {
                                        "type": "number",
                                        "example": 1500000,
                                        "nullable": true
                                    },
                                    "client_project": {
                                        "type": "string",
                                        "example": "Saray Towers",
                                        "nullable": true
                                    },
                                    "client_developer": {
                                        "type": "string",
                                        "example": "WR",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Moved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "lead": {
                                                    "$ref": "#/components/schemas/Lead"
                                                },
                                                "latest_path": {
                                                    "$ref": "#/components/schemas/TicketPath"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found or not visible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiError": {
                "description": "Standard API error response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "code": {
                        "type": "string",
                        "example": "INVALID_CREDENTIALS"
                    },
                    "message": {
                        "type": "string",
                        "example": "Invalid credentials"
                    },
                    "data": {
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 134
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 7
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    }
                },
                "type": "object"
            },
...
            "Lead": {
                "description": "Lead summary object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "full_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "00971501234567"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "status": {
                        "type": "string",
                        "example": "follow-up"
                    },
                    "assigned_to": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 12
                            },
                            "name": {
                                "type": "string",
                                "example": "Sales User"
                            }
                        },
                        "type": "object"
                    },
                    "last_follow_up": {
                        "type": "string",
                        "example": "Called the client and scheduled a meeting",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-10T09:30:00Z"
                    }
                },
                "type": "object"
            },
            "LeadPermissions": {
                "properties": {
                    "can_view": {
                        "type": "boolean",
                        "example": true
                    },
                    "can_view_paths": {
                        "type": "boolean",
                        "example": true
                    },
                    "can_view_all_paths": {
                        "type": "boolean",
                        "example": false
                    },
                    "can_update_status": {
                        "type": "boolean",
                        "example": true
                    },
                    "can_add_follow_up": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "LastFollowUp": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 555
                    },
                    "comment": {
                        "type": "string",
                        "example": "Call client tomorrow"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "next_user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 13
                            },
                            "name": {
                                "type": "string",
                                "example": "Tony"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "next_status": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "name": {
                                "type": "string",
                                "example": "Follow-up"
                            },
                            "slug": {
                                "type": "string",
                                "example": "follow-up"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object",
                "nullable": true
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}